hybrid-curiosity-arena/requirements.txt
# Core ML/RL Dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Web Interface
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# Visualization
matplotlib>=3.5.0
plotly>=5.15.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Computer Vision (for environment rendering)
opencv-python>=4.8.0
pillow>=9.5.0

# Data Analysis
pandas>=1.5.0
scikit-learn>=1.1.0

# Development
pytest>=7.0.0
black>=22.0.0
